requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.0.284
chromadb==0.4.15
openai==1.3.5 
supabase==2.0.3
sqlalchemy==2.0.23
pydantic==1.10.13
python-jose[cryptography]==3.3.0
unstructured==0.10.30
Pillow==10.1.0
pytesseract==0.3.10
tiktoken==0.5.2